Modified the code to support a register file with synchronous read. This allows users to choose between different register file architectures, including synchronous RAMs. This may help reduce overall gate count in both ASIC and FPGA architectures.
The change in the synchronous read from the register file required to change from
a single-cycle CPU to a multi-cycle CPU. For now the CPU is two cycle, such that
the cycles are named `q1` and `q4` to alias with the cycle naming in the legacy
PIC architecture. Cycles `q2` and `q3` are not present, but could be added if
desired. Note that the CPU is really multi-cycle (i.e. one instruction taking
multiple cycles with no instruction overlapping) rather than pipelined. This is
to minimize the extent of changes to the legacy code.

The timing is such that the `q1` cycle is used for register file read and `q4`
is used for register file write. `q1` is similarly used for all external "reads"
(i.e. program memory read or expansion interface read). Port A is also sampled
in `q1` so that its value is available for `q4` where the instruction execution
happens.

As noted above, the program memory timing is as follows. The address is valid
in `q1` and new instruction shall be available in `q4`. If the program memory
is synchronous, it may either be clocked at the same clock as the CPU, or at
half the frequency if its clock is properly aligned with the CPU one.